requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
asyncpg==0.30.0
beanie==1.29.0
dependency-injector==4.45.
alembic==1.14.1
argon2-cffi==23.1.0
pyjwt==2.9
pydantic[email]
ollama
aioredis==2.0.1 
redis==5.2.1